Store temporary files in a temporary tmpfs
This avoids leaking /tmp/ts_isolate.* files.

The other approach is to just unlink the file in /tmp after doing the
bind mount, which works, but it leaves the bind mount in a state where
you can't bind mount on top of it - i.e., that you can't nest ts_isolate
inside ts_isolate. We rely on nesting internally in a couple of obscure
cases.

In order to allow us to actually create files in the tmpfs, move the
uid_map/gid_map logic before profile creation, so that we have a mapped
UID/GID in the tmpfs's user namespace when we try to call open(). This
doesn't break anything - I forgot that capabilities over the new user
namespace remain effective until exec or explicitly dropped.

(Our internal code used to unlink, but it also ignored error returns
from mount, and the unlink call was accidentally lost in a refactoring
to add better error checking.)